add mandataris-edit-prompt component
