+++ app/components/mandatarissen/mandataris-edit-prompt.hbs
@@ -1,0 +1,49 @@
+{{#unless this.editMode}}
+  <div class="au-o-grid au-o-grid--small">
+    <div class="au-o-grid__item au-u-1-2">
+      <AuContent @skin="small">
+        <AuHeading @level="4" @skin="5">Wijziging huidige situatie</AuHeading>
+        <p>Indien de mandataris wordt vervangen, of de huidige
+          <strong>status, fractie, rangorde of bevoegdheid</strong>
+          verandert:
+        </p>
+        <ul>
+          <li>beëindig dit mandaat door de einddatum in te vullen;</li>
+          <li><strong>start een nieuw mandaat</strong>
+            met de gewijzigde situatie.</li>
+        </ul>
+        <AuButton {{on "click" this.terminate}}>Wijzig einde mandaat</AuButton>
+      </AuContent>
+    </div>
+    <div class="au-o-grid__item au-u-1-2">
+      <AuContent @skin="small">
+        <AuHeading @level="4" @skin="5">Correctie</AuHeading>
+        <p>Indien een mandaat
+          <strong>fouten</strong>
+          bevat, en niet goed ingevoerd werd in het verleden,
+          <strong>corrigeer</strong>
+          de fouten.
+        </p>
+        <AuButton @skin="secondary" {{on "click" this.correct}}>Corrigeer fouten</AuButton>
+      </AuContent>
+    </div>
+  </div>
+{{/unless}}
+<div class="au-o-region-large">
+  {{#if this.isTerminating}}
+    <Form::Instance
+      @instanceId={{@mandataris.id}}
+      @form={{@mandatarisEindeEditForm}}
+      @onSave={{this.onSave}}
+      @onCancel={{this.cancel}}
+    />
+  {{else if this.isCorrecting}}
+    <Form::Instance
+      @instanceId={{@mandataris.id}}
+      @form={{@mandatarisEditForm}}
+      @onSave={{this.onSave}}
+      @onCancel={{this.cancel}}
+      @buildMetaTtl={{this.buildMetaTtl}}
+    />
+  {{/if}}
+</div>
