+++ app/templates/mandatarissen/mandataris.hbs
@@ -12,59 +12,17 @@
       Add mandaat card here.
     </Tmp::Todo>
 
-    {{#unless this.editMode}}
-      <div class="au-o-grid au-o-grid--small">
-        <div class="au-o-grid__item au-u-1-2">
-          <AuContent @skin="small">
-            <AuHeading @level="4" @skin="5">Wijziging huidige situatie</AuHeading>
-            <p>Indien de mandataris wordt vervangen, of de huidige
-              <strong>status, fractie, rangorde of bevoegdheid</strong>
-              verandert:
-            </p>
-            <ul>
-              <li>beëindig dit mandaat door de einddatum in te vullen;</li>
-              <li><strong>start een nieuw mandaat</strong>
-                met de gewijzigde situatie.</li>
-            </ul>
-            <AuButton {{on "click" this.terminate}}>Wijzig einde mandaat</AuButton>
-          </AuContent>
-        </div>
-        <div class="au-o-grid__item au-u-1-2">
-          <AuContent @skin="small">
-            <AuHeading @level="4" @skin="5">Correctie</AuHeading>
-            <p>Indien een mandaat
-              <strong>fouten</strong>
-              bevat, en niet goed ingevoerd werd in het verleden,
-              <strong>corrigeer</strong>
-              de fouten.
-            </p>
-            <AuButton @skin="secondary" {{on "click" this.correct}}>Corrigeer
-              fouten</AuButton>
-          </AuContent>
-        </div>
-      </div>
-    {{/unless}}
-    <div class="au-o-region-large">
-      {{#if this.isTerminating}}
-        <Form::Instance
-          @instanceId={{@model.mandataris.id}}
-          @form={{@model.mandatarisEindeEditForm}}
-          @onSave={{this.onSave}}
-          @onCancel={{this.cancel}}
-        />
-      {{else if this.isCorrecting}}
-        <Form::Instance
-          @instanceId={{@model.mandataris.id}}
-          @form={{@model.mandatarisEditForm}}
-          @onSave={{this.onSave}}
-          @onCancel={{this.cancel}}
-          @buildMetaTtl={{this.buildMetaTtl}}
-        />
-      {{/if}}
+    {{#if this.editing}}
+      <Mandatarissen::MandatarisEditPrompt
+        class="au-c-collapsible au-u-margin-top-small"
+        @mandataris={{@model.mandataris}}
+        @mandatarisEindeEditForm={{@model.mandatarisEindeEditForm}}
+        @mandatarisEditForm={{@model.mandatarisEditForm}}
+      />
+    {{/if}}
 
-      <Tmp::Todo>
-        Add history here.
-      </Tmp::Todo>
-    </div>
+    <Tmp::Todo>
+      Add history here.
+    </Tmp::Todo>
   </div>
 </div>
